Workaround: If all joints are commanded with zero velocity, the robot started to relax, which was undesired.
Actually the problem was not concerned with the velocity, but more with commanding the current positions as new ones.
In that case the robot relaxed. As a workaround the robot still moves to the bounds now, but with a velocity of 1[tick].
Due to friction the robot stops at current position.